페이지 복구 오류 해결, 페이지 완전 제거 기능 추가

+++ withSpace/src/main/java/hansung/cse/withSpace/service/PageService.java
@@ -7,14 +7,21 @@ import hansung.cse.withSpace.domain.space.TrashCan;
 import hansung.cse.withSpace.exception.page.PageDeletionNotAllowedException;
 import hansung.cse.withSpace.exception.page.PageNotFoundException;
 import hansung.cse.withSpace.exception.page.PageNotInSpaceException;
+import hansung.cse.withSpace.exception.page.PageRestoreNotCurrentPageIdException;
 import hansung.cse.withSpace.repository.PageRepository;
+import hansung.cse.withSpace.repository.SpaceRepository;
 import hansung.cse.withSpace.requestdto.space.page.PageCreateRequestDto;
 import hansung.cse.withSpace.requestdto.space.page.PageUpdateContentRequestDto;
 import hansung.cse.withSpace.requestdto.space.page.PageUpdateTitleRequestDto;
 import hansung.cse.withSpace.responsedto.space.page.PageHierarchyDto;
 import hansung.cse.withSpace.responsedto.space.page.PageTrashCanDto;
+import jakarta.persistence.EntityManager;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.Session;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
@@ -28,6 +35,7 @@ import java.util.Optional;
 public class PageService {
 
     private final PageRepository pageRepository;
+    private final SpaceRepository spaceRepository;
     private final SpaceService spaceService;
 
     public Page findOne(Long pageId) {
@@ -113,13 +121,21 @@ public class PageService {
         TrashCan trashCan = space.getTrashCan();
         List<Page> pageTrashCanList = trashCan.getPageList();
 
+
+
+        if (page.getParentPage() != null) { //부모 페이지가 있다면 연결을 끊어줘야함
+            page.removeParentPageRelationWhenThrow();
+        }
+
         PageTrashCanDto pageTrashCanDto = new PageTrashCanDto(page);
-        System.out.println(pageTrashCanDto+"------------------------------------------");
+
         page.putTrashCan(trashCan);//본인 쓰레기통에 넣고
+
 
         for (Page chlidPage : page.getChildPages()) {
             chlidPage.putTrashCan(trashCan); //자식페이지들도 쓰레기통에
         }
+
         return pageTrashCanDto;
 
     }
@@ -140,24 +156,50 @@ public class PageService {
             space.setTopLevelPageCount(space.getTopLevelPageCount() - 1);
         }
 
-        return moveToTrashCan(page,space);
+        PageTrashCanDto pageTrashCanDto = moveToTrashCan(page, space);
+
+        System.out.println();
+
+
+        return pageTrashCanDto;
 
     }
 
     @Transactional
-    public void restorePageAndChildren(Long pageId, Long spaceId) { //페이지 복구
+    public void restorePageAndChildren(Long pageId, Long spaceId, Long currentPageId) { //페이지 복구
         Page page = findOne(pageId);
         Space space = spaceService.findOne(spaceId);
         TrashCan trashCan = page.getTrashCan();
 
-        page.outTrashCan(trashCan); // 쓰레기통에서 페이지 제거
-        //page.makeRelationPageSpace(page, space); //스페이스와의 연관관계 다시 연결
+        if (page.getParentPage() != null) {
+            //부모가 있는 페이지인데 특정 자식페이지만 살릴라면 기존 부모페이지와의 연결 끊어주고
+            //그리고 현재 보고있는 페이지를 부모 페이지로 해줌
+            //이 경우 currentPageId가 반드시 넘어와야함
+            if (currentPageId == null) {
+                throw new PageRestoreNotCurrentPageIdException("현재 보고 있는 페이지의 Id 정보 - currentPageId 누락");
+            }
+
+            Page currentPage = findOne(currentPageId);
+            page.removeParentPageRelationWhenRestore(currentPage);
+
+            //page.restoreParentPageRelation(currentPage);
+
+        }
+
+        if (page.getBeforeParentId() != null) { //끊어진 부모페이지와의 관계 이어줌
+            Page beforeParent = findOne(page.getBeforeParentId());
+            page.outTrashCan(trashCan, beforeParent); // 쓰레기통에서 페이지 제거
+        }else{
+            page.outTrashCan(trashCan); // 쓰레기통에서 페이지 제거
+        }
 
         //자식 페이지들도 마찬가지
         List<Page> childPages = page.getChildPages();
         for (Page childPage : childPages) {
             childPage.outTrashCan(trashCan); // 쓰레기통에서 페이지 제거
         }
+
+        System.out.println(page.getSpace().getId()+"-----------------");
     }
 
 
@@ -166,20 +208,12 @@ public class PageService {
     public void deletePage(Long pageId) {
         Page page = findOne(pageId);
 
-//        Space space = page.getSpace();
-//
-//        if (space.getTopLevelPageCount() == 1 && page.getParentPage() == null ) {
-//            //스페이스에 최상위 페이지가 하나 + 삭제하려는 페이지가 또 제일 최상단 부모페이지면 삭제 불가
-//
-//            throw new PageDeletionNotAllowedException("최상위 페이지가 하나밖에 없는 경우에는 삭제가 불가능합니다.");
-//
-//        }
-//
-//        if (page.getParentPage() == null) { //최상위 페이지인경우
-//            space.setTopLevelPageCount(space.getTopLevelPageCount() - 1);
-//        }
+        if (page.getTrashCan() == null) { //페이지가 쓰레기통에 있는지 검사
+            throw new PageDeletionNotAllowedException("페이지가 쓰레기통에 없습니다.");
+        }
 
-
+        List<Page> childPages = page.getChildPages();
+        pageRepository.deleteAll(childPages);
         pageRepository.delete(page);
     }
